+++ yash-env/CHANGELOG.md
@@ -20,6 +20,11 @@ A _private dependency_ is used internally and not visible to downstream users.
   variable name. This allows modules to check variable names without directly
   depending on the `yash-syntax` crate.
 
+### Changed
+
+- `impl PartialEq for function::Function`: Now compares the function bodies
+  using pointer equality instead of deep equality.
+
 ### Removed
 
 - `parser::is_name`: This re-export of `yash_syntax::parser::lex::is_name` has